Extract DNS and reporting functions from main file
